refactor(converters, services): Simplify logging and navigation services
- Removed the `LogFormattingConverter` class to streamline log entry formatting, retaining only the `LogHighlightConverter` for backward compatibility.
- Updated the `LoggingService` to improve event handling by renaming the private event for clarity and consistency.
- Removed the unused `DefaultViewModel` class from the `NavigationService`, enhancing code cleanliness and maintainability.
