+++ src/Services/DefaultViewModel.cs
@@ -1,0 +1,12 @@
+// <copyright file="DefaultViewModel.cs" company="PlaceholderCompany">
+// Copyright (c) PlaceholderCompany. All rights reserved.
+// </copyright>
+
+namespace AppIntBlockerGUI.Services
+{
+    using CommunityToolkit.Mvvm.ComponentModel;
+
+    public class DefaultViewModel : ObservableObject
+    {
+    }
+}
